[ALPS07225892] imgsys: fix coverity
Fix coverity

MTK-Commit-Id: 208e664d163d6ca6110cb0ad068a95739aca6b5b

Feature: [Android Default] Camera Application Basic Functions
CR-Id: ALPS07225892

+++ drivers/media/platform/mtk-isp/mtk-hcp/mtk-hcp.c
@@ -102,7 +102,7 @@ struct packet {
 
 struct ctrl_data {
 	uint32_t id;
-	uintptr_t value;
+	uint64_t value;
 } __packed;
 
 #define HCP_INIT                _IOWR('H', 0, struct share_buf)
